Runs directly from home directory now
Renames the bash files to include the leading dot, and removes the
install-bashfiles script.

Adds a .gitignore file, by default, ignores everything except files
that are explicitly excluded (by using a leading bang).

The intent is that the project should live directly in the home
directory. Git does not allow cloning into a non-empty directory, so to
set this up initially, you need to clone to a new directory and move
the contents into $HOME.