Read GITHUB_TOKEN from ~/.github.token.
Thie environment variable is used by the 'hub' script, and possibly others,
as a substitute for the github.token git configuration parameter.

This lets me leave it out of .gitconfig (which is stored, publicly, in this
repository). The ~/.github.token file needs to be locally recreated on each
computer.

Note that this is a temporary measure: token authentication is deprecated in
github and slated for removal.